Add project update form
