+++ src/css/projects/projects-page.css
@@ -10,7 +10,7 @@
 
 .project-group-container {
     width: 100%;
-    margin-bottom: 2rem;
+    margin-bottom: 5rem;
 }
 
 .projectGrp{
